requirements.txt
pandas
wordcloud
statsmodels.api
plotly.express
numpy
matplotlib.pyplot
sklearn.cluster